RW-2718 begin react conversion
